requirements.txt
flask
psycopg2-binary
email_validator
Flask-SQLAlchemy
Flask_WTF
pandas
matplotlib